lm_nav/requirements.txt
networkx==2.8.4
h5py==3.7.0
matplotlib==3.5.2
numpy==1.23.0
pillow
pyyaml
utm==0.7.0
openai==0.20.0
jupyter==1.0.0
opencv-python==4.6.0.66
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
ftfy==6.1.1
regex==2022.6.2
tqdm==4.64.0
git+https://github.com/openai/CLIP.git
spacy==3.3.1
